# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Event Photo App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: number;
  /** Short shareable code for this event */
  code: string;
  name: string;
  /** Whether this event has an admin passcode set */
  hasAdminPasscode: boolean;
  photoCount: number;
  createdAt: string;
}

export interface EventInput {
  /**
     * Event name
     * @minLength 1
     */
  name: string;
  /**
     * Optional admin passcode for managing the event
     * @nullable
     */
  adminPasscode?: string | null;
}

export interface Photo {
  id: number;
  eventId: number;
  /** Storage path to the photo */
  objectPath: string;
  uploadedAt: string;
}

export interface PhotoInput {
  /** Storage path returned from the upload URL endpoint */
  objectPath: string;
}

export interface AdminVerifyInput {
  passcode: string;
}

export interface AdminVerifyResponse {
  valid: boolean;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}