# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Photo App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Event management
  - name: photos
    description: Photo management
  - name: admin
    description: Admin operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /events:
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /events/{code}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by short code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /events/{code}/photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List all photos for an event
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: addPhoto
      tags: [photos]
      summary: Register an uploaded photo to an event
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Photo registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /events/{code}/photos/{photoId}:
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo (admin only)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Photo deleted
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /events/{code}/admin/verify:
    post:
      operationId: verifyAdminPasscode
      tags: [admin]
      summary: Verify admin passcode for an event
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVerifyResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: File not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve public object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Event:
      type: object
      required: [id, code, name, hasAdminPasscode, photoCount, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
          description: Short shareable code for this event
        name:
          type: string
        hasAdminPasscode:
          type: boolean
          description: Whether this event has an admin passcode set
        photoCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    EventInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: Event name
        adminPasscode:
          type: ["string", "null"]
          description: Optional admin passcode for managing the event

    Photo:
      type: object
      required: [id, eventId, objectPath, uploadedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        objectPath:
          type: string
          description: Storage path to the photo
        uploadedAt:
          type: string
          format: date-time

    PhotoInput:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
          description: Storage path returned from the upload URL endpoint

    AdminVerifyInput:
      type: object
      required: [passcode]
      properties:
        passcode:
          type: string

    AdminVerifyResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string