# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Event Photo App API
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  /** Short shareable code for this event */
  code: string;
  name: string;
  /** Whether this event has an admin passcode set */
  hasAdminPasscode: boolean;
  photoCount: number;
  createdAt: Date;
}