# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Event Photo App API
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  /**
     * Event name
     * @minLength 1
     */
  name: string;
  /**
     * Optional admin passcode for managing the event
     * @nullable
     */
  adminPasscode?: string | null;
}