# lib/api-zod/src/generated/types/photo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Event Photo App API
 * OpenAPI spec version: 0.1.0
 */

export interface Photo {
  id: number;
  eventId: number;
  /** Storage path to the photo */
  objectPath: string;
  uploadedAt: Date;
}