# lib/api-zod/src/generated/types/photoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Event Photo App API
 * OpenAPI spec version: 0.1.0
 */

export interface PhotoInput {
  /** Storage path returned from the upload URL endpoint */
  objectPath: string;
}