# lib/api-zod/src/generated/types/uploadUrlRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Event Photo App API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}